+++ src/Jieba/Serializer/SerializerFactory.php
@@ -12,8 +12,15 @@ use Jieba\Exception;
 class SerializerFactory
 {
     const DEFAULT = 'default';
+    const BSON    = 'bson';
     const JSON    = 'json';
     const MSGPACK = 'msgpack';
+
+    const EXTENSIONS = [
+        self::BSON    => 'bson',
+        self::JSON    => 'json',
+        self::MSGPACK => 'mp',
+    ];
 
     /**
      * @var SerializerInterface
@@ -25,17 +32,20 @@ class SerializerFactory
      * @return SerializerInterface
      * @throws Exception
      */
-    public static function getSerializer(string $type = self::DEFAULT)
+    public static function getSerializer(string $type = self::DEFAULT): SerializerInterface
     {
         if (self::DEFAULT == $type) {
             if (!isset(self::$serializer)) {
-                self::$serializer = self::getSerializer((extension_loaded('msgpack') ? self::MSGPACK : self::JSON));
+                self::$serializer = self::getSerializer(self::getAllAvailableTypes()[0]);
             }
 
             return self::$serializer;
         }
 
         switch ($type) {
+            case self::BSON:
+                return new Bson();
+                break;
             case self::JSON:
                 return new Json();
                 break;
@@ -60,12 +70,19 @@ class SerializerFactory
     }
 
     /**
+     * Return a list of available serializer types ordered by performance. The first one is the fastest.
+     *
      * @return array
+     * @todo add benchmark on dictionary files.
      */
     public static function getAllAvailableTypes(): array
     {
-        $serializers = [self::JSON];
+        $serializers = [];
 
+        if (extension_loaded('mongodb')) {
+            $serializers[] = self::BSON;
+        }
+        $serializers[] = self::JSON;
         if (extension_loaded('msgpack')) {
             $serializers[] = self::MSGPACK;
         }